% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summary.R
\name{my_summary}
\alias{my_summary}
\title{Summary method for my_lm objects}
\usage{
my_summary(object, ...)
}
\arguments{
\item{object}{An object of class "my_lm"}

\item{...}{Not used}
}
\value{
Invisibly returns a list with coefficient table, etc.
}
\description{
Summary method for my_lm objects
}
