% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fast linear model using RcppArmadillo}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{An object of class "formula"}

\item{data}{A data.frame containing the variables in the model}
}
\value{
An object of class "my_lm" similar to lm() output
}
\description{
A simplified version of lm() that uses RcppArmadillo for the core
linear algebra. Supports formulas of the form y ~ x1 + x2 + ...
}
\examples{
\dontrun{
  set.seed(1)
  x <- rnorm(100)
  y <- 1 + 2 * x + rnorm(100, sd = 0.5)
  df <- data.frame(x = x, y = y)
  fit <- my_lm(y ~ x, df)
  summary(fit)
}
}
