% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm_original.R
\name{my_lm_original}
\alias{my_lm_original}
\title{A pure R implementation without Rcpp}
\usage{
my_lm_original(formula, data)
}
\arguments{
\item{formula}{An object of class "formula"}

\item{data}{A data.frame containing the variables in the model}
}
\value{
An object of class "my_lm" similar to lm() output
}
\description{
A pure R implementation without Rcpp
}
